Add statistical outlier removal, voxel downsampling of combined pointcloud and ability to fill small holes in the mesh.
The downsampling is done to have a consistant density in the pointcloud
used in normal estimation and reconstruction. This makes it possbile to
reason about the number of nearest neighbours used in those steps.
e.g. if we know the points are in a 1cm grid using 100 neighbours
corresponds to a circle with 11cm diameter (ignoring the third axis).